+++ lib/services/subgraph.service.ts
@@ -17,8 +17,6 @@ export interface Vote {
     title: string;
   };
   blockNumber: number;
-  nounHolderLikes: number;
-  nonNounHolderLikes: number;
 }
 
 export interface Proposal {
