+++ lib/types/VoteWithLikes.ts
@@ -1,0 +1,6 @@
+import { Vote } from '../services/subgraph.service';
+
+export interface VoteWithLikes extends Vote {
+  nounHolderLikes: number;
+  nonNounHolderLikes: number;
+}
